bug fixes and added fMRI preprocessing options

+++ steps/feat.bash
@@ -4,8 +4,7 @@
 #SBATCH --mem=2G                      #max memory per node
 #SBATCH --partition=luna-cpu-short    #using luna short queue
 #SBATCH --cpus-per-task=1      	      #max CPU cores per process
-#SBATCH --time=0:20:00                #time limit (H:MM:SS)
-#SBATCH --nice=2000                   #allow other priority jobs to go first
+#SBATCH --time=0:30:00                #time limit (H:MM:SS)
 #SBATCH --qos=anw-cpu                 #use anw-cpu's
 #SBATCH --output=logs/slurm-%x.%j.out
 
@@ -39,10 +38,17 @@ restingstate=$3
 subfolder=$4/files
 delete_vols=$5
 skip_slice_timing=$6
-FULLID_folder=$7
+slice_order_file=$7
+FULLID_folder=$8
+dummy_reg=$9
 outputdir=${PWD}/func/${FULLID_folder}
 restingstate_json=${restingstate%%.nii.gz}.json
-[ -f ${outputdir}/SynBOLD_DisCo/output/BOLD_u.nii.gz ] && restingstate=${outputdir}/SynBOLD_DisCo/output/BOLD_u.nii.gz
+
+[ ${dummy_reg} -eq 1 ] && fslroi $restingstate ${outputdir}/fmri_dummy.nii.gz 0 1
+
+if [ -f ${outputdir}/SynBOLD_DisCo/output/BOLD_u.nii.gz ]; then 
+    restingstate=${outputdir}/SynBOLD_DisCo/output/BOLD_u.nii.gz
+fi
 
 #Check if script has already been completed
 [ -d ${outputdir}/fmri.feat ] && exit 0
@@ -65,7 +71,7 @@ numvoxels=$(fslhd ${restingstate} | grep ^dim[1/2/3/4] | awk '{print $2}' | past
 cp ${subfolder}/feat_settings.fsf ${outputdir}/feat_settings.fsf &&\
 
 #adjust the file to use subject specific paths
-sed -i 's|ANATOMICAL_scan|'${anatomical_brain}'|g' ${outputdir}/feat_settings.fsf &&\
+sed -i 's|ANATOMICAL_scan|'${anatomical_brain}'|' ${outputdir}/feat_settings.fsf &&\
 sed -i 's|FMRI_TR|'$(printf ${TR})'|' ${outputdir}/feat_settings.fsf &&\
 sed -i 's|N_VOL|'$(printf ${volnum})'|' ${outputdir}/feat_settings.fsf &&\
 sed -i 's|N_DELETE|'$(printf ${delete_vols})'|' ${outputdir}/feat_settings.fsf &&\
@@ -73,12 +79,23 @@ sed -i 's|RESTINGSTATE|'${restingstate}'|' ${outputdir}/feat_settings.fsf &&\
 sed -i 's|NUMVOXELS|'${numvoxels}'|' ${outputdir}/feat_settings.fsf &&\
 sed -i 's|STANDARDBRAIN|"'${FSLDIR}'/data/standard/MNI152_T1_2mm_brain"|' ${outputdir}/feat_settings.fsf &&\
 sed -i 's|OUTPUTDIR|"'${outputdir}'/fmri.feat"|' ${outputdir}/feat_settings.fsf &&\
-if [ ${skip_slice_timing} -eq 0 ];then
+[ -f ${outputdir}/SynBOLD_DisCo/output/BOLD_u.nii.gz ] && sed -i 's|set fmri(mc) 1|set fmri(mc) 0|' ${outputdir}/feat_settings.fsf &&\
+
+if [[ ${slice_order_file} != 'NA' ]] && [ -f ${slice_order_file} ]; then
+    cp ${slice_order_file} ${outputdir}/slice_order.txt
+    sed -i 's|fmri(st) 0|fmri(st) 3|' ${outputdir}/feat_settings.fsf &&\
+    sed -i 's|SLICE_ORDER_FILE|"'${outputdir}'/slice_order.txt"|' ${outputdir}/feat_settings.fsf
+elif [ ${skip_slice_timing} -eq 0 ];then
     jq '.SliceTiming[]' ${restingstate_json} | tr -d '[,]'| awk 'NF' > ${outputdir}/slicetime_tmp.txt &&\
     cat -n ${outputdir}/slicetime_tmp.txt | sort -nsk2,2 | awk '{print $1}' > ${outputdir}/slice_order.txt &&\
     rm ${outputdir}/slicetime_tmp.txt &&\
     sed -i 's|fmri(st) 0|fmri(st) 3|' ${outputdir}/feat_settings.fsf &&\
     sed -i 's|SLICE_ORDER_FILE|"'${outputdir}'/slice_order.txt"|' ${outputdir}/feat_settings.fsf
+fi
+
+if [ ${dummy_reg} -eq 1 ]; then
+    sed -i 's|fmri(reginitial_highres_yn) 0|fmri(reginitial_highres_yn) 1|' ${outputdir}/feat_settings.fsf &&\
+    sed -i 's|INITIAL_HIGHRES|'${outputdir}'/fmri_dummy.nii.gz|' ${outputdir}/feat_settings.fsf
 fi
 
 #copy anatomical to same folder as anatomical (required for feat to work)
@@ -89,9 +106,11 @@ fi
 
 ##run FEAT
 printf "Performing FEAT...\n" &&\
-feat ${outputdir}/feat_settings.fsf &&\
+feat ${outputdir}/feat_settings.fsf
+[ ! -f ${outputdir}/fmri.feat/filtered_func_data.nii.gz ] && exit 1
 rm ${outputdir}/feat_settings.fsf &&\
 mv ${outputdir}/slice_order.txt ${outputdir}/fmri.feat/slice_order.txt &&\
+[ -f ${outputdir}/fmri_dummy.nii.gz ] && rm ${outputdir}/fmri_dummy.nii.gz
 
 #change permissions (bug)
 chmod -R u+rwx ${outputdir}/fmri.feat
